[Backport MR-9648] BUG: error for non-fatal xml issue
As the original author indicated [1], this was meant to be a warning,
but the use of `vtkErrorMacro` indicates and error even when
the read is successful.

For example, reading an empty `vtkPolyData` results
in an error message even though the object is correctly read.

This leads to backwards incompatible behavior as discussed in [2].

[1] https://gitlab.kitware.com/vtk/vtk/-/commit/7f768465185b4dbe04ab23de0a7a6034db4a2e61

[2] https://github.com/Slicer/Slicer/issues/6447