[Backport] Remove unnecessary "No input data" error message in vtkTransformPolyDataFilter
vtkTransformPolyDataFilter is often used in pipelines where it is completely valid to have empty input.
For example, when slicing through a mesh with a plane, the output may be empty if the plane does not happen
to intersect any part of the mesh. Logging an error pollutes the error log and if performed too frequently
then it may slow down an application.

The workarounds that applications applied was either add an error observer to the vtkTransformPolyDataFilter
(but that may suppress valid errors, too) or force an Update() on the input producing filter and prevent the
vtkTransformPolyDataFilter from running (but that may be very complex in a long pipeline and may lead to
unnecessary Update() calls, which slow down the application).
See for example in 3D Slicer: https://github.com/Slicer/Slicer/pull/7212